Add parser tests [WIK-724] (#5)
Add a test that the parser function is usable by testing against mediawiki.org
and checking that the version returned starts with `1.`. We cannot easily test
that the output matches a pattern, so just test the start of the version.
`functionhooks` are not used to ensure that the extension is installed because
if they are used, when the extension is not installed the tests are skipped and
pass rather than fail as desired. Thus, just do not check that the extension is
loaded, and if it is not the test will fail.

WIK-724