Do not crash if addwiki/mediawiki-api-base is missing
Instead of crashing with a fatal error and a stack trace, check if one of the
classes that we need is available, using that class as a proxy for the overall
package. If not, uses of the parser function return a nice error message but
the parsing does not fail and the rest of the page can be shown.

WIK-1250